Change mapId type to uint32 in LoginVerifyWorld packet
mapId is now serialized as a 32-bit unsigned integer (uint32)
instead of a 64-bit unsigned integer (uint64) when sending the
LoginVerifyWorld packet to the client. This ensures correct
data size and alignment for the protocol.

+++ src/shared/game/game.h
@@ -9,7 +9,7 @@
 
 namespace mmo
 {
-	typedef uint64 MapId;
+	typedef uint32 MapId;
 
 	typedef uuids::uuid InstanceId;
